AC Power Calculator Ver1
Calculates AC energy based on BUT, wattage, size, and type. Whole lotta error checking so a bit of boilerplate with the questions but the research for proper conversion rates was way more tedious than I could've ever imagined because kWh is the only standardized measurement for price checking but BTU is used for all ACs. Just going to leave this here as is just to get a commit down. CSV has a ton of data that I've collected, should be enough for awhile.